Add headers to the api-date.js file
