+++ src/api-personal-key.js
@@ -1,0 +1,3 @@
+const exports = require("webpack");
+
+const WEATHER_API_KEY = "79cb4dadbbmsh995b2f6bed49e17p17ce0fjsn5277305d5ee8";
